+++ src/components/SocialWorker/ElderyRegister/HealthConditionSection.tsx
@@ -3,7 +3,7 @@ import {
   SectionWrapper,
   Title,
   TitleWrapper,
-} from '@/components/SocialWorker/ElderyCreate/Section.styles';
+} from '@/components/SocialWorker/ElderyRegister/Section.styles';
 
 interface HealthConditionSectionProps {
   healthCondition: string;
