+++ src/components/SocialWorker/ElderyRegister/InmateSection.tsx
@@ -6,7 +6,7 @@ import {
   SectionWrapper,
   Title,
   TitleWrapper,
-} from '@/components/SocialWorker/ElderyCreate/Section.styles';
+} from '@/components/SocialWorker/ElderyRegister/Section.styles';
 
 interface InmateSectionProps {
   inmate: '있음' | '없음' | '';
