--- src/data/CareTypeList.tsx
@@ -1,28 +1,0 @@
-export const CareTypeList = {
-  식사보조: [
-    '스스로 식사 가능',
-    '식사 차려드리기',
-    '죽, 반찬 등 요리 필요',
-    '경관식 보조',
-  ],
-  이동보조: [
-    '스스로 거동 가능',
-    '이동시 부축 도움',
-    '휠체어 이동 보조',
-    '거동 불가',
-  ],
-  배변보조: [
-    '스스로 배변 가능',
-    '가끔 대소변 실수시 도움',
-    '기저귀 케어 필요',
-    '유치도뇨/방광루/장루관리',
-  ],
-  일상생활: [
-    '청소, 빨래 보조',
-    '목욕 보조',
-    '인지자극 활동',
-    '말벗 등 정서지원',
-    '산책, 간단한 운동',
-  ],
-  질병보조: ['병원 동행', '복약지도', '재활프로그램', '인지치료'],
-};
